updated @grafana/create-plugin
