+++ .config/.cprc.json
@@ -1,3 +1,4 @@
 {
-  "version": "5.25.9"
+  "version": "7.6.0",
+  "features": {}
 }
