+++ docker-compose.yaml
@@ -1,19 +1,14 @@
 services:
   grafana:
-    user: root
-    container_name: 'grafana-timestream-datasource'
-
     build:
-      context: ./.config
       args:
-        grafana_image: ${GRAFANA_IMAGE:-grafana-enterprise}
         grafana_version: ${GRAFANA_VERSION:-latest}
     ports:
       - 3000:3000/tcp
-    volumes:
-      - ./dist:/var/lib/grafana/plugins/grafana-timestream-datasource
-      - ./provisioning:/etc/grafana/provisioning
 
     environment:
       ACCESS_KEY: ${ACCESS_KEY}
       SECRET_KEY: ${SECRET_KEY}
+    extends:
+      file: .config/docker-compose-base.yaml
+      service: grafana
