+++ package.json
@@ -6,7 +6,7 @@
     "build": "webpack -c ./.config/webpack/webpack.config.ts --env production",
     "dev": "webpack -w -c ./.config/webpack/webpack.config.ts --env development",
     "e2e": "playwright test",
-    "lint": "eslint --cache .",
+    "lint": "eslint --cache --ignore-path ./.gitignore --ext .js,.jsx,.ts,.tsx .",
     "lint:fix": "yarn run lint --fix && prettier --write --list-different .",
     "server": "docker compose up --build",
     "sign": "npx --yes @grafana/sign-plugin@latest",
@@ -22,11 +22,12 @@
   "license": "Apache-2.0",
   "dependencies": {
     "@emotion/css": "11.13.5",
-    "@grafana/data": "^12.1.0",
+    "@grafana/data": "^12.2.0",
+    "@grafana/i18n": "^12.2.0",
     "@grafana/plugin-ui": "^0.13.0",
-    "@grafana/runtime": "^12.1.0",
-    "@grafana/schema": "^12.1.0",
-    "@grafana/ui": "^12.1.0",
+    "@grafana/runtime": "^12.2.0",
+    "@grafana/schema": "^12.2.0",
+    "@grafana/ui": "^12.2.0",
     "react": "18.3.1",
     "react-dom": "18.3.1",
     "react-router-dom": "^6.22.0",
@@ -51,8 +52,9 @@
     "@testing-library/react": "16.3.2",
     "@types/jest": "30.0.0",
     "@types/node": "24.12.0",
+    "@types/react": "^18.3.0",
+    "@types/react-dom": "^18.3.0",
     "@types/semver": "7.7.1",
-    "@types/testing-library__jest-dom": "5.14.9",
     "@typescript-eslint/eslint-plugin": "8.57.1",
     "@typescript-eslint/parser": "8.57.1",
     "copy-webpack-plugin": "14.0.0",
@@ -61,7 +63,6 @@
     "dotenv": "17.3.1",
     "eslint": "9.39.4",
     "eslint-config-prettier": "10.1.8",
-    "eslint-plugin-deprecation": "3.0.0",
     "eslint-plugin-jsdoc": "62.8.0",
     "eslint-plugin-prettier": "5.5.5",
     "eslint-plugin-react": "7.37.5",
@@ -70,26 +71,26 @@
     "fork-ts-checker-webpack-plugin": "9.1.0",
     "glob": "13.0.6",
     "identity-obj-proxy": "3.0.0",
-    "imports-loader": "5.0.0",
+    "imports-loader": "^5.0.0",
     "jest": "30.3.0",
     "jest-environment-jsdom": "30.3.0",
     "prettier": "3.8.1",
     "react-select-event": "5.5.1",
-    "replace-in-file-webpack-plugin": "1.0.6",
+    "replace-in-file-webpack-plugin": "^1.0.6",
     "sass": "1.98.0",
     "sass-loader": "16.0.7",
     "semver": "7.7.4",
     "style-loader": "4.0.0",
     "swc-loader": "0.2.7",
     "terser-webpack-plugin": "5.4.0",
-    "ts-node": "10.9.2",
+    "ts-node": "^10.9.2",
     "tsconfig-paths": "4.2.0",
     "typescript": "5.9.3",
     "webpack": "5.105.4",
     "webpack-cli": "7.0.2",
-    "webpack-livereload-plugin": "3.0.2",
-    "webpack-subresource-integrity": "5.1.0",
-    "webpack-virtual-modules": "0.6.2"
+    "webpack-livereload-plugin": "^3.0.2",
+    "webpack-subresource-integrity": "^5.1.0",
+    "webpack-virtual-modules": "^0.6.2"
   },
   "engines": {
     "node": ">=18"
